pyproject.toml
[project]
name = "test-login"
version = "0.1.0"
description = ""
authors = [
    {name = "Aman Kadam",email = "126617073+AmanKadam-16@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite (>=0.21.0,<0.22.0)",
    "mailjet-rest (>=1.5.1,<2.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)"
]

[tool.poetry]
packages = [{include = "test_login", from = "src"}]

[tool.poetry.dependencies]
fastapi = "^0.115.0"
uvicorn = "^0.23.0"
sqlalchemy = "^2.0"
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = "^2.8.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
api = "test_login.main:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
